expand buttons count to 16

+++ include/zmk/hid-joystick/joystick.h
@@ -6,5 +6,5 @@
 
 #pragma once
 
-typedef uint8_t zmk_joystick_button_flags_t;
+typedef uint16_t zmk_joystick_button_flags_t;
 typedef uint16_t zmk_joystick_button_t;
